STYLE: Canvas setting code arrange
