+++ src/features/canvas/Canvas.js
@@ -2,7 +2,8 @@ import './Canvas.css';
 import CanavasLayer from './CanvasLayer';
 import DrawBtnLayer from './DrawBtnLayer';
 import CanvasRotated from './CanvasRotated';
-import { useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
+import MainCanvasDraw from '../../controls/mainCanvasDraw';
 
 export default function Canavas({ isRotated }) {
     return (
@@ -13,11 +14,35 @@ export default function Canavas({ isRotated }) {
 }
 
 function IsRotatedCanvas({ isRotated }) {
+    const CANVAS = { width: 1190, height: 1684 };
 
     const canvasRef = {
         canvasMainRef : useRef(null),
         canvasFlagRef : useRef(null)
     };
+
+    const [ctxFlag, setCtxFlag] = useState();
+    const [ctxMain, setCtxMain] = useState();
+
+    function drawDefaultSetting(ctxMain) {
+        MainCanvasDraw.defaultSet(ctxMain);
+    }
+
+    useEffect(() => {
+        const canvasMain = canvasRef.canvasMainRef.current;
+        const ctxMain = canvasMain.getContext('2d');
+        canvasMain.width = CANVAS.width;
+        canvasMain.height = CANVAS.height;
+
+        const canvasFlag = canvasRef.canvasFlagRef.current;
+        const ctxFlag = canvasFlag.getContext('2d');
+        canvasFlag.width = CANVAS.width;
+        canvasFlag.height = CANVAS.height;
+
+        setCtxFlag(ctxFlag);
+        setCtxMain(ctxMain);
+        drawDefaultSetting(ctxMain);
+    });
 
     const [listFlag, setListFlag] = useState([]);
 
@@ -25,8 +50,8 @@ function IsRotatedCanvas({ isRotated }) {
         return <CanvasRotated />;
     } else {
         return <>
-            <DrawBtnLayer listFlag={listFlag} ref={canvasRef}/>
-            <CanavasLayer listFlag={listFlag} setListFlag={setListFlag} ref={canvasRef}/>
+            <DrawBtnLayer listFlag={listFlag} ctxFlag={ctxFlag} ctxMain={ctxMain}/>
+            <CanavasLayer listFlag={listFlag} setListFlag={setListFlag} ctxFlag={ctxFlag} ref={canvasRef}/>
         </>;
     }
 }
