+++ src/features/canvas/DrawBtnLayer.js
@@ -1,30 +1,16 @@
-import { forwardRef, useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { InputNum, DropDownKey, InputTempo, BtnInfo, DropDownSongform, UploadImage, DropDownFlag } from './DrawBtns';
 import MainCanvasDraw from '../../controls/mainCanvasDraw';
 import './DrawBtnLayer.css';
 import FlagCanvasDraw from '../../controls/flagCanvasDraw';
 
-const DrawBtnLayer = forwardRef(function DrawBtn({listFlag}, ref) {
-    const { canvasMainRef, canvasFlagRef } = ref;
+export default function DrawBtn({ listFlag, ctxMain, ctxFlag }) {
     const imageRef = useRef();
-
-    const [ctxMain, setCtxMain] = useState();
-    const [ctxFlag, setCtxFlag] = useState();
 
     const listSongform = [];
 
     const mainCanvasDraw = new MainCanvasDraw(ctxMain);
     const flagCanvasDraw = new FlagCanvasDraw(ctxFlag);
-
-    useEffect(() => {
-        const canvasMain = canvasMainRef.current;
-        const contextMain = canvasMain.getContext('2d');
-        setCtxMain(contextMain);
-
-        const canvasFlag = canvasFlagRef.current;
-        const contextFlag = canvasFlag.getContext('2d');
-        setCtxFlag(contextFlag);
-    }, [])
 
     function btnControls(trigger, value = '', event = '') {
         switch (trigger) {
@@ -74,9 +60,7 @@ const DrawBtnLayer = forwardRef(function DrawBtn({listFlag}, ref) {
             <BtnInfo btnControls={btnControls} />
             <DropDownSongform btnControls={btnControls} />
             <UploadImage ref={imageRef} btnControls={btnControls} />
-            <DropDownFlag btnControls={btnControls}/>
+            <DropDownFlag btnControls={btnControls} />
         </div>
     );
-});
-
-export default DrawBtnLayer;
+}
